@@ -1,106 +1,160 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Request;
+
 use Cassandra\Protocol\Frame;
 use Cassandra\Type;
+use Cassandra\Value;
 
-class Request implements Frame{
+abstract class Request implements Frame
+{
+    public const CONSISTENCY_ANY = 0x0000;
+    public const CONSISTENCY_ONE = 0x0001;
+    public const CONSISTENCY_TWO = 0x0002;
+    public const CONSISTENCY_THREE = 0x0003;
+    public const CONSISTENCY_QUORUM = 0x0004;
+    public const CONSISTENCY_ALL = 0x0005;
+    public const CONSISTENCY_LOCAL_QUORUM = 0x0006;
+    public const CONSISTENCY_EACH_QUORUM = 0x0007;
+    public const CONSISTENCY_SERIAL = 0x0008;
+    public const CONSISTENCY_LOCAL_SERIAL = 0x0009;
+    public const CONSISTENCY_LOCAL_ONE = 0x000A;
 
-    const CONSISTENCY_ANY = 0x0000;
-    const CONSISTENCY_ONE = 0x0001;
-    const CONSISTENCY_TWO = 0x0002;
-    const CONSISTENCY_THREE = 0x0003;
-    const CONSISTENCY_QUORUM = 0x0004;
-    const CONSISTENCY_ALL = 0x0005;
-    const CONSISTENCY_LOCAL_QUORUM = 0x0006;
-    const CONSISTENCY_EACH_QUORUM = 0x0007;
-    const CONSISTENCY_SERIAL = 0x0008;
-    const CONSISTENCY_LOCAL_SERIAL = 0x0009;
-    const CONSISTENCY_LOCAL_ONE = 0x000A;
-    
+    protected int $version = 3;
+
+    protected int $opcode;
+
+    protected int $stream = 0;
+
+    protected int $flags = 0;
+
     /**
-     * @var int
+     * @var ?array<string,string> $payload
      */
-    protected $version = 0x03;
-    
+    protected ?array $payload = null;
+
     /**
-     * @var int
+     * @param ?array<string,string> $payload
      */
-    protected $opcode;
-    
-    /**
-     * @var int
-     */
-    protected $stream = 0;
-    
-    /**
-     * @var int
-     */
-    protected $flags = 0;
-    
-    /**
-     * @param int $opcode
-     * @param int $stream
-     * @param int $flags
-     */
-    public function __construct($opcode, $stream = 0, $flags = 0) {
+    public function __construct(int $opcode, int $stream = 0, int $flags = 0, array $payload = null, int $version = 3)
+    {
         $this->opcode = $opcode;
         $this->stream = $stream;
         $this->flags = $flags;
+        $this->payload = $payload;
+        $this->version = $version;
     }
-        
-    public function getVersion(){
-        return $this->version;
-    }
-    
-    public function getFlags(){
-        return $this->flags;
-    }
-    
-    public function getStream(){
-        return $this->stream;
-    }
-    
-    public function getOpcode(){
+
+    public function getOpcode(): int
+    {
         return $this->opcode;
     }
-    
-    public function getBody(){
+
+    public function getStream(): int
+    {
+        return $this->stream;
+    }
+
+    public function getFlags(): int
+    {
+        return $this->flags;
+    }
+
+    /**
+     * @return ?array<string,string>
+     */
+    public function getPayload(): ?array
+    {
+        return $this->payload;
+    }
+
+    public function getVersion(): int
+    {
+        return $this->version;
+    }
+
+    public function getBody(): string
+    {
         return '';
     }
-    
-    public function setStream($stream){
+
+    public function setStream(int $stream): void
+    {
         $this->stream = $stream;
     }
-    
+
+    public function setFlags(int $flags): void
+    {
+        $this->flags = $flags;
+    }
+
+    public function enableTracing(): void
+    {
+        $this->flags |= self::FLAG_TRACING;
+    }
+
     /**
-     * @return string
+     * @param array<string,string> $payload
      */
-    public function __toString(){
+    public function setPayload(array $payload): void
+    {
+        $this->payload = $payload;
+        $this->flags |= self::FLAG_CUSTOM_PAYLOAD;
+    }
+
+    public function setVersion(int $version): void
+    {
+        $this->version = $version;
+    }
+
+    public function __toString(): string
+    {
         $body = $this->getBody();
+
+        if ($this->flags & self::FLAG_CUSTOM_PAYLOAD) {
+            if ($this->payload === null) {
+                $this->flags &= ~self::FLAG_CUSTOM_PAYLOAD;
+            } else {
+                $payloadData = pack('n', count($this->payload));
+
+                foreach ($this->payload as $key => $val) {
+                    $payloadData .= pack('n', strlen($key)) . $key;
+                    $payloadData .= pack('N', strlen($val)) . $val;
+                }
+
+                $body = $payloadData . $body;
+            }
+        }
+
         return pack(
-                'CCnCN',
-                $this->version,
-                $this->flags,
-                $this->stream,
-                $this->opcode,
-                strlen($body)
+            'CCnCN',
+            $this->version,
+            $this->flags,
+            $this->stream,
+            $this->opcode,
+            strlen($body)
         ) . $body;
     }
-    
+
     /**
-     * 
-     * @param array $values
-     * @throws Type\Exception
-     * @return string
+     * @param array<mixed> $values
+     *
+     * @throws \Cassandra\Type\Exception
      */
-    public static function valuesBinary(array $values, $namesForValues = false) {
+    public static function valuesBinary(array $values, bool $namesForValues = false): string
+    {
         $valuesBinary = pack('n', count($values));
-        
-        $index = 0;
-        foreach($values as $name => $value) {
+
+        /** @var mixed $value */
+        foreach ($values as $name => $value) {
             switch (true) {
                 case $value instanceof Type\Base:
                     $binary = $value->getBinary();
+                    break;
+                case $value instanceof Value\NotSet:
+                    $binary = $value;
                     break;
                 case $value === null:
                     $binary = null;
@@ -118,68 +172,85 @@ class Request implements Frame{
                     throw new Type\Exception('Unknown type.');
             }
 
-            if ($namesForValues){
-                $valuesBinary .= pack('n', strlen($name)) . strtolower($name);
-            }
-            else{
+            if ($namesForValues) {
+                if (is_string($name)) {
+                    $valuesBinary .= pack('n', strlen($name)) . strtolower($name);
+                } else {
+                    throw new Type\Exception('$values should be an associative array given, sequential array given. Or you can set "names_for_values" option to false.');
+                }
+            } elseif (is_string($name)) {
                 /**
-                 * @see https://github.com/duoshuo/php-cassandra/issues/29
-                 */
-                if ($index++ !== $name)
-                    throw new Type\Exception('$values should be an sequential array, associative array given.  Or you can set "names_for_values" option to true.');
+                * @see https://github.com/duoshuo/php-cassandra/issues/29
+                */
+                throw new Type\Exception('$values should be an sequential array, associative array given. Or you can set "names_for_values" option to true.');
             }
 
-            $valuesBinary .= $binary === null
-                ? "\xff\xff\xff\xff"
-                : pack('N', strlen($binary)) . $binary;
+            if ($binary === null) {
+                $valuesBinary .= "\xff\xff\xff\xff";
+            } elseif ($binary instanceof Value\NotSet) {
+                $valuesBinary .= "\xff\xff\xff\xfe";
+            } else {
+                $valuesBinary .= pack('N', strlen($binary)) . $binary;
+            }
         }
-        
+
         return $valuesBinary;
     }
-    
+
     /**
-     * 
-     * @param array $values
-     * @param array $columns
-     * @return array
+     * @param array<mixed> $values
+     * @param array<array{name: string, type: int|array<mixed>}> $columns
+     * @return array<mixed>
+     *
+     * @throws \Cassandra\Type\Exception
      */
-    public static function strictTypeValues(array $values, array $columns) {
+    public static function strictTypeValues(array $values, array $columns): array
+    {
         $strictTypeValues = [];
-        foreach($columns as $index => $column) {
+        foreach ($columns as $index => $column) {
             $key = array_key_exists($column['name'], $values) ? $column['name'] : $index;
-            
-            if (!isset($values[$key])){
+
+            if (!isset($values[$key])) {
                 $strictTypeValues[$key] = null;
-            }
-            elseif($values[$key] instanceof Type\Base){
+            } elseif ($values[$key] instanceof Type\Base) {
                 $strictTypeValues[$key] = $values[$key];
-            }
-            else{
+            } else {
                 $strictTypeValues[$key] = Type\Base::getTypeObject($column['type'], $values[$key]);
             }
         }
-        
+
         return $strictTypeValues;
     }
-    
+
     /**
-     * 
-     * @param int $consistency
-     * @param array $values
-     * @param array $options
-     * @return string
+     * @param array<mixed> $values
+     * @param array{
+     *  names_for_values?: bool,
+     *  skip_metadata?: bool,
+     *  page_size?: int,
+     *  paging_state?: string,
+     *  serial_consistency?: int,
+     *  default_timestamp?: int,
+     *  keyspace?: string,
+     *  now_in_seconds?: int,
+     * } $options
+     *
+     * @throws \Cassandra\Type\Exception
+     * @throws \Cassandra\Request\Exception
      */
-    public static function queryParameters($consistency, array $values = [], array $options = []){
+    public static function queryParameters(int $consistency, array $values = [], array $options = [], int $version = 3): string
+    {
         $flags = 0;
         $optional = '';
-        
-        if (!empty($values)) {
+
+        if ($values) {
             $flags |= Query::FLAG_VALUES;
             $optional .= Request::valuesBinary($values, !empty($options['names_for_values']));
         }
 
-        if (!empty($options['skip_metadata']))
+        if (!empty($options['skip_metadata'])) {
             $flags |= Query::FLAG_SKIP_METADATA;
+        }
 
         if (isset($options['page_size'])) {
             $flags |= Query::FLAG_PAGE_SIZE;
@@ -201,9 +272,32 @@ class Request implements Frame{
             $optional .= Type\Bigint::binary($options['default_timestamp']);
         }
 
-        if (!empty($options['names_for_values']))
-            $flags |= Query::FLAG_WITH_NAMES_FOR_VALUES;
+        if (isset($options['keyspace'])) {
+            if ($version >= 5) {
+                $flags |= Query::FLAG_WITH_KEYSPACE;
+                $optional .= pack('n', strlen($options['keyspace'])) . $options['keyspace'];
+            } else {
+                throw new Exception('Option "keyspace" not supported by server');
+            }
+        }
 
-        return pack('n', $consistency) . pack('C', $flags) . $optional;
+        if (isset($options['now_in_seconds'])) {
+            if ($version >= 5) {
+                $flags |= Query::FLAG_WITH_NOW_IN_SECONDS;
+                $optional .= pack('N', $options['now_in_seconds']);
+            } else {
+                throw new Exception('Option "now_in_seconds" not supported by server');
+            }
+        }
+
+        if (!empty($options['names_for_values'])) {
+            $flags |= Query::FLAG_WITH_NAMES_FOR_VALUES;
+        }
+
+        if ($version < 5) {
+            return pack('n', $consistency) . pack('C', $flags) . $optional;
+        } else {
+            return pack('n', $consistency) . pack('N', $flags) . $optional;
+        }
     }
 }
