@@ -1,61 +1,75 @@
 <?php
+
+declare(strict_types=1);
+
 /* This file can be used instead of install with composer.
- * Just include "require 'PATH/php-cassandra/php-cassandra.php';" to your code (where PATH is path to php-cassandra folder).
+ * Just include "require __DIR__  . '/PATH/php-cassandra/php-cassandra.php';" to your code (where PATH is path to php-cassandra folder).
  */
 
-require 'src/Exception.php';
+require __DIR__  . '/src/Exception.php';
 
-require 'src/Type/Base.php';
-require 'src/Type/Varchar.php';
-require 'src/Type/Ascii.php';
-require 'src/Type/Bigint.php';
-require 'src/Type/Blob.php';
-require 'src/Type/Boolean.php';
-require 'src/Type/CollectionList.php';
-require 'src/Type/CollectionMap.php';
-require 'src/Type/CollectionSet.php';
-require 'src/Type/Counter.php';
-require 'src/Type/Custom.php';
-require 'src/Type/Decimal.php';
-require 'src/Type/Double.php';
-require 'src/Type/Exception.php';
-require 'src/Type/PhpFloat.php';
-require 'src/Type/Inet.php';
-require 'src/Type/PhpInt.php';
-require 'src/Type/Timestamp.php';
-require 'src/Type/Uuid.php';
-require 'src/Type/Timeuuid.php';
-require 'src/Type/Tuple.php';
-require 'src/Type/UDT.php';
-require 'src/Type/Varint.php';
+require __DIR__  . '/src/Type/Exception.php';
+require __DIR__  . '/src/Type/Base.php';
+require __DIR__  . '/src/Type/Varchar.php';
+require __DIR__  . '/src/Type/Ascii.php';
+require __DIR__  . '/src/Type/Bigint.php';
+require __DIR__  . '/src/Type/Blob.php';
+require __DIR__  . '/src/Type/Boolean.php';
+require __DIR__  . '/src/Type/CollectionList.php';
+require __DIR__  . '/src/Type/CollectionMap.php';
+require __DIR__  . '/src/Type/CollectionSet.php';
+require __DIR__  . '/src/Type/Counter.php';
+require __DIR__  . '/src/Type/Custom.php';
+require __DIR__  . '/src/Type/Date.php';
+require __DIR__  . '/src/Type/Decimal.php';
+require __DIR__  . '/src/Type/Double.php';
+require __DIR__  . '/src/Type/Duration.php';
+require __DIR__  . '/src/Type/PhpFloat.php';
+require __DIR__  . '/src/Type/Inet.php';
+require __DIR__  . '/src/Type/PhpInt.php';
+require __DIR__  . '/src/Type/Smallint.php';
+require __DIR__  . '/src/Type/Timestamp.php';
+require __DIR__  . '/src/Type/Uuid.php';
+require __DIR__  . '/src/Type/Timeuuid.php';
+require __DIR__  . '/src/Type/Tinyint.php';
+require __DIR__  . '/src/Type/Tuple.php';
+require __DIR__  . '/src/Type/UDT.php';
+require __DIR__  . '/src/Type/Varint.php';
 
-require 'src/Protocol/Frame.php';
+require __DIR__  . '/src/Value/NotSet.php';
 
-require 'src/Connection/SocketException.php';
-require 'src/Connection/Socket.php';
-require 'src/Connection/StreamException.php';
-require 'src/Connection/Stream.php';
+require __DIR__  . '/src/Protocol/Frame.php';
 
-require 'src/Request/Request.php';
-require 'src/Request/AuthResponse.php';
-require 'src/Request/Batch.php';
-require 'src/Request/Execute.php';
-require 'src/Request/Options.php';
-require 'src/Request/Prepare.php';
-require 'src/Request/Query.php';
-require 'src/Request/Register.php';
-require 'src/Request/Startup.php';
+require __DIR__  . '/src/Connection/NodeException.php';
+require __DIR__  . '/src/Connection/SocketException.php';
+require __DIR__  . '/src/Connection/StreamException.php';
+require __DIR__  . '/src/Connection/Node.php';
+require __DIR__  . '/src/Connection/NodeImplementation.php';
+require __DIR__  . '/src/Connection/Socket.php';
+require __DIR__  . '/src/Connection/Stream.php';
 
-require 'src/Response/StreamReader.php';
-require 'src/Response/Response.php';
-require 'src/Response/Authenticate.php';
-require 'src/Response/AuthSuccess.php';
-require 'src/Response/Error.php';
-require 'src/Response/Event.php';
-require 'src/Response/Exception.php';
-require 'src/Response/Ready.php';
-require 'src/Response/Result.php';
-require 'src/Response/Supported.php';
+require __DIR__  . '/src/Request/Exception.php';
+require __DIR__  . '/src/Request/Request.php';
+require __DIR__  . '/src/Request/AuthResponse.php';
+require __DIR__  . '/src/Request/Batch.php';
+require __DIR__  . '/src/Request/Execute.php';
+require __DIR__  . '/src/Request/Options.php';
+require __DIR__  . '/src/Request/Prepare.php';
+require __DIR__  . '/src/Request/Query.php';
+require __DIR__  . '/src/Request/Register.php';
+require __DIR__  . '/src/Request/Startup.php';
 
-require 'src/Connection.php';
-require 'src/Statement.php';
+require __DIR__  . '/src/Response/Exception.php';
+require __DIR__  . '/src/Response/StreamReader.php';
+require __DIR__  . '/src/Response/Response.php';
+require __DIR__  . '/src/Response/ResultIterator.php';
+require __DIR__  . '/src/Response/Authenticate.php';
+require __DIR__  . '/src/Response/AuthSuccess.php';
+require __DIR__  . '/src/Response/Error.php';
+require __DIR__  . '/src/Response/Event.php';
+require __DIR__  . '/src/Response/Ready.php';
+require __DIR__  . '/src/Response/Result.php';
+require __DIR__  . '/src/Response/Supported.php';
+
+require __DIR__  . '/src/Connection.php';
+require __DIR__  . '/src/Statement.php';
