@@ -1,9 +1,16 @@
 <?php
-namespace Cassandra\Response;
-use Cassandra\Protocol\Frame;
 
-class Authenticate extends Response {
-	public function getData(){
-		return unpack('n', $this->getBody())[1];
-	}
+declare(strict_types=1);
+
+namespace Cassandra\Response;
+
+class Authenticate extends Response
+{
+    /**
+     * @throws \Cassandra\Response\Exception
+     */
+    public function getData(): string
+    {
+        return $this->_stream->readString();
+    }
 }
