@@ -1,5 +1,9 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Ascii extends Varchar{
+class Ascii extends Varchar
+{
 }
