@@ -1,5 +1,9 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class CollectionSet extends CollectionList{
+class CollectionSet extends CollectionList
+{
 }
