@@ -1,5 +1,9 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Counter extends Bigint{
+class Counter extends Bigint
+{
 }
