@@ -1,45 +1,114 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Decimal extends Base{
+class Decimal extends Base
+{
+    protected ?string $_value = null;
 
     /**
-     * @param string $value
-     * @throws Exception
+     * @throws \Cassandra\Type\Exception
      */
-    public function __construct($value = null){
-        if (!is_numeric($value))
+    public function __construct(?string $value = null)
+    {
+        if (!is_numeric($value)) {
             throw new Exception('Incoming value must be numeric string.');
+        }
 
         $this->_value = $value;
     }
-    
-    public static function binary($value){
-            $pos = strpos($value, '.');
-            $scaleLen = $pos === false ? 0 : strlen($value) - $pos - 1;
-            $value *= pow(10, $scaleLen);
-            $higher = ($value & 0xffffffff00000000) >> 32;
-            $lower = $value & 0x00000000ffffffff;
-            $binary = pack('NNN', $scaleLen, $higher, $lower);
+
+    /**
+     * @param mixed $value
+     * @param null|int|array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
+     */
+    protected static function create(mixed $value, null|int|array $definition): self
+    {
+        if ($value !== null && !is_string($value)) {
+            throw new Exception('Invalid value type');
+        }
+
+        return new self($value);
+    }
+
+    /**
+     * @throws \Cassandra\Type\Exception
+     */
+    public function binaryOfValue(): string
+    {
+        if ($this->_value === null) {
+            throw new Exception('value is null');
+        }
+
+        return static::binary($this->_value);
+    }
+
+    /**
+     * @throws \Cassandra\Type\Exception
+     */
+    public function parseValue(): ?string
+    {
+        if ($this->_value === null && $this->_binary !== null) {
+            $this->_value = static::parse($this->_binary);
+        }
+
+        return $this->_value;
+    }
+
+    public function __toString(): string
+    {
+        return (string) $this->_value;
+    }
+
+    public static function binary(string $value): string
+    {
+        $pos = strpos($value, '.');
+        $scaleLen = $pos === false ? 0 : strlen($value) - $pos - 1;
+        if ($scaleLen) {
+            $numericValue = (float)$value * pow(10, $scaleLen);
+        } else {
+            $numericValue = (int)$value;
+        }
+        $higher = ($numericValue & 0xffffffff00000000) >> 32;
+        $lower = $numericValue & 0x00000000ffffffff;
+        $binary = pack('NNN', $scaleLen, $higher, $lower);
         return $binary;
     }
-    
+
     /**
-     * @return string
+     * @param null|int|array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
      */
-    public static function parse($binary){
-        $unpacked = unpack('N1scale/C*', $this->_binary);
+    public static function parse(string $binary, null|int|array $definition = null): string
+    {
+        $length = strlen($binary);
+
+        /**
+         * @var false|array<int> $unpacked
+         */
+        $unpacked = unpack('N1scale/C*', $binary);
+        if ($unpacked === false) {
+            throw new Exception('Cannot unpack binary.');
+        }
+
         $valueByteLen = $length - 4;
         $value = 0;
-        for ($i = 1; $i <= $valueByteLen; ++$i)
+        for ($i = 1; $i <= $valueByteLen; ++$i) {
             $value |= $unpacked[$i] << (($valueByteLen - $i) * 8);
-        $shift = (\PHP_INT_SIZE - $valueByteLen) * 8;
-        $value = $value << $shift >> $shift;
-        if ($unpacked['scale'] === 0)
-            return (string) $value;
-        elseif (strlen($value) > $unpacked['scale'])
+        }
+        $shift = (PHP_INT_SIZE - $valueByteLen) * 8;
+        $value = (string) ($value << $shift >> $shift);
+        if ($unpacked['scale'] === 0) {
+            return $value;
+        } elseif (strlen($value) > $unpacked['scale']) {
             return substr($value, 0, -$unpacked['scale']) . '.' . substr($value, -$unpacked['scale']);
-        else
+        } else {
             return $value >= 0 ? sprintf("0.%0$unpacked[scale]d", $value) : sprintf("-0.%0$unpacked[scale]d", -$value);
+        }
     }
 }
