@@ -1,4 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Exception extends \Cassandra\Exception {}
+use Cassandra\Exception as CassandraException;
+
+class Exception extends CassandraException
+{
+}
