@@ -1,26 +1,83 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class PhpFloat extends Base{
-    /**
-     * @param double $value
-     * @throws Exception
-     */
-    public function __construct($value = null){
-        if ($value === null)
-            return;
-        
-        if (!is_double($value))
-            throw new Exception('Incoming value must be type of double.');
-    
+class PhpFloat extends Base
+{
+    protected ?float $_value = null;
+
+    public function __construct(?float $value = null)
+    {
         $this->_value = $value;
     }
-    
-    public static function binary($value){
-        return strrev(pack('f', $value));
+
+    /**
+     * @param mixed $value
+     * @param null|int|array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
+     */
+    protected static function create(mixed $value, null|int|array $definition): self
+    {
+        if ($value !== null && !is_float($value)) {
+            throw new Exception('Invalid value type');
+        }
+
+        return new self($value);
     }
-    
-    public static function parse($binary){
-        return unpack('f', strrev($binary))[1];
+
+    /**
+     * @throws \Cassandra\Type\Exception
+     */
+    public function binaryOfValue(): string
+    {
+        if ($this->_value === null) {
+            throw new Exception('value is null');
+        }
+
+        return static::binary($this->_value);
+    }
+
+    /**
+     * @throws \Cassandra\Type\Exception
+     */
+    public function parseValue(): ?float
+    {
+        if ($this->_value === null && $this->_binary !== null) {
+            $this->_value = static::parse($this->_binary);
+        }
+
+        return $this->_value;
+    }
+
+    public function __toString(): string
+    {
+        return (string) $this->_value;
+    }
+
+    public static function binary(float $value): string
+    {
+        return strrev(pack('G', $value));
+    }
+
+    /**
+     * @param null|int|array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
+     */
+    public static function parse(string $binary, null|int|array $definition = null): float
+    {
+        /**
+         * @var false|array<float> $unpacked
+         */
+        $unpacked = unpack('G', strrev($binary));
+
+        if ($unpacked === false) {
+            throw new Exception('Cannot unpack binary.');
+        }
+
+        return $unpacked[1];
     }
 }
