@@ -1,5 +1,9 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Timestamp extends Bigint{
+class Timestamp extends Bigint
+{
 }
