@@ -1,5 +1,9 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Timeuuid extends Uuid{
+class Timeuuid extends Uuid
+{
 }
