@@ -1,50 +1,121 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Cassandra\Type;
 
-class Tuple extends Base{
+use Cassandra\Response\StreamReader;
+
+class Tuple extends Base
+{
+    /**
+     * @var array<int|array<mixed>> $_definition
+     */
+    protected array $_definition;
 
     /**
-     * @param array $value
-     * @param array $definition
-     * @throws Exception
+     * @var ?array<mixed> $_value
      */
-    public function __construct($value, array $definition){
-        $this->_definition = $definition;
-        
-        if ($value === null)
-            return;
-        
-        if (!is_array($value))
-            throw new Exception('Incoming value must be type of array.');
+    protected ?array $_value = null;
 
+    /**
+     * @param ?array<mixed> $value
+     * @param array<int|array<mixed>> $definition
+     */
+    public function __construct(?array $value, array $definition)
+    {
+        $this->_definition = $definition;
         $this->_value = $value;
     }
 
-    public static function binary($value, array $definition){
+    /**
+     * @param mixed $value
+     * @param null|int|array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
+     */
+    protected static function create(mixed $value, null|int|array $definition): self
+    {
+        if ($value !== null && !is_array($value)) {
+            throw new Exception('Invalid value type');
+        }
+
+        if (!is_array($definition)) {
+            throw new Exception('Invalid definition');
+        }
+
+        return new self($value, $definition);
+    }
+
+    /**
+     * @throws \Cassandra\Type\Exception
+     */
+    public function binaryOfValue(): string
+    {
+        if ($this->_value === null) {
+            throw new Exception('value is null');
+        }
+
+        return static::binary($this->_value, $this->_definition);
+    }
+
+    /**
+     * @return ?array<mixed> $_value
+     *
+     * @throws \Cassandra\Response\Exception
+     * @throws \Cassandra\Type\Exception
+     */
+    public function parseValue(): ?array
+    {
+        if ($this->_value === null && $this->_binary !== null) {
+            $this->_value = static::parse($this->_binary, $this->_definition);
+        }
+
+        return $this->_value;
+    }
+
+    public function __toString(): string
+    {
+        return (string) json_encode($this->_value);
+    }
+
+    /**
+     * @param array<mixed> $value
+     * @param array<int|array<mixed>> $definition
+     *
+     * @throws \Cassandra\Type\Exception
+     */
+    public static function binary(array $value, array $definition): string
+    {
         $binary = '';
         foreach ($definition as $key => $type) {
             if ($value[$key] === null) {
                 $binary .= "\xff\xff\xff\xff";
-            }
-            else {
+            } else {
                 $valueBinary = $value[$key] instanceof Base
                     ? $value[$key]->getBinary()
                     : Base::getBinaryByType($type, $value[$key]);
-                
+
                 $binary .= pack('N', strlen($valueBinary)) . $valueBinary;
             }
         }
-        
+
         return $binary;
     }
-    
+
     /**
-     * 
-     * @param string $binary
-     * @param array $definition
-     * @return array
+     * @param null|int|array<int|array<mixed>> $definition
+     * @return array<mixed>
+     *
+     * @throws \Cassandra\Response\Exception
+     * @throws \Cassandra\Type\Exception
      */
-    public static function parse($binary, array $definition){
-        return (new \Cassandra\Response\StreamReader($binary))->readTuple($definition);
+    public static function parse(string $binary, null|int|array $definition = null): array
+    {
+        if (!is_array($definition)) {
+            throw new Exception('invalid Tuple definition');
+        }
+
+        return (new StreamReader($binary))->readTuple($definition);
     }
 }
